Added PMS5003 polling intervals
Previously, the PMS5003 sensor could be set in passive mode with
only one measurement being taken before the sensor is placed back
to sleep. However, measurements shortly after the warmup period
may be inaccurate and it may be desirable to do a longer sample
period to get a better value.

This commit allows the PMS5003 sensor to be polled for an arbitrary
duration within the wake interval. The polling interval can be set
as a second argument to the sensor (eg. Sensor18 2,polling-interval)

To keep backwards compatibility with the old behavior, the sensor18
command will also take the wake interval like before with the
polling interval set to 1 for a single measurement.